requirements.txt
# Requirements.txt file is used to specify what dependencies are
# needed to for the package to run rather than for deployment of a tested set of
# packages.

ftputil
dnspython
#pwntools
psutil
#watchdog
requests
python-registry
ftplib
-----------------------
for astro
-----------------------
opencv-python-headless
numpy
astroalign
scikit-image
argparse
